Ditch admin_users

+++ app/controllers/phrases_controller.rb
@@ -1,15 +1,17 @@
 class PhrasesController < ApplicationController
-  before_filter :authorize_user
+  before_filter :authenticate_user!
+  before_filter :load_audio_video
 
   def index
-    @phrases = Phrase.by_audio_video(params[:audio_video_id])
+    @phrases = current_user.phrases_by_audio_video(@audio_video)
     render :json => @phrases
   end
 
   def create
-    audio_video = AudioVideo.find(audio_video_id)
-    phrase = Phrase.new(phrase_params)
-    phrase.audio_video_id = audio_video.id
+    phrase = Phrase.new(phrase_params).tap do |phrase|
+      phrase.audio_video = @audio_video
+      phrase.user = current_user
+    end
 
     if phrase.save
       render :json => phrase
@@ -20,10 +22,18 @@ class PhrasesController < ApplicationController
 
   def destroy
     @phrase = Phrase.find(phrase_id)
-    render :json => @phrase.destroy
+    if @phrase.user == current_user
+      render :json => @phrase.destroy
+    else
+      render :text => 'Forbiden fot this user', :status => 403
+    end
   end
 
   private
+
+  def load_audio_video
+    @audio_video = AudioVideo.find audio_video_id
+  end
 
   def phrase_id
     params.required(:id)
